fix: made changes according to feedback
