+++ src/components/Avatar/Avatar.js
@@ -1,0 +1,4 @@
+const Avatar = (props) => {
+  return <img className={props.className} src={props.src} alt={props.alt} />;
+};
+export default Avatar;
