+++ src/components/UserItem/UserItem.module.css
@@ -11,6 +11,13 @@
   flex-direction: row;
   align-items: center;
 }
+.avatar {
+  border-radius: 50%;
+  width: 2.8rem;
+  margin-right: 1rem;
+  height: 2.8rem;
+}
+
 .name_email {
   flex: 2;
 }
@@ -19,12 +26,6 @@
 }
 .email {
   color: #313131;
-}
-.avatar {
-  border-radius: 50%;
-  width: 2.8rem;
-  margin-right: 1rem;
-  height: 2.8rem;
 }
 
 .icon {
