+++ src/store/userReducer.js
@@ -1,0 +1,112 @@
+export const actionTypes = {
+  DELETE_USER: "delete",
+  SHOW_USER: "userdetail",
+};
+
+const initialValues = {
+  data: [
+    {
+      id: 1,
+      email: "george.bluth@reqres.in",
+      first_name: "George",
+      last_name: "Bluth",
+      avatar: "https://reqres.in/img/faces/1-image.jpg",
+    },
+    {
+      id: 2,
+      email: "janet.weaver@reqres.in",
+      first_name: "Janet",
+      last_name: "Weaver",
+      avatar: "https://reqres.in/img/faces/2-image.jpg",
+    },
+    {
+      id: 3,
+      email: "emma.wong@reqres.in",
+      first_name: "Emma",
+      last_name: "Wong",
+      avatar: "https://reqres.in/img/faces/3-image.jpg",
+    },
+    {
+      id: 4,
+      email: "eve.holt@reqres.in",
+      first_name: "Eve",
+      last_name: "Holt",
+      avatar: "https://reqres.in/img/faces/4-image.jpg",
+    },
+    {
+      id: 5,
+      email: "charles.morris@reqres.in",
+      first_name: "Charles",
+      last_name: "Morris",
+      avatar: "https://reqres.in/img/faces/5-image.jpg",
+    },
+    {
+      id: 6,
+      email: "tracey.ramos@reqres.in",
+      first_name: "Tracey",
+      last_name: "Ramos",
+      avatar: "https://reqres.in/img/faces/6-image.jpg",
+    },
+    {
+      id: 7,
+      email: "michael.lawson@reqres.in",
+      first_name: "Michael",
+      last_name: "Lawson",
+      avatar: "https://reqres.in/img/faces/7-image.jpg",
+    },
+    {
+      id: 8,
+      email: "lindsay.ferguson@reqres.in",
+      first_name: "Lindsay",
+      last_name: "Ferguson",
+      avatar: "https://reqres.in/img/faces/8-image.jpg",
+    },
+    {
+      id: 9,
+      email: "tobias.funke@reqres.in",
+      first_name: "Tobias",
+      last_name: "Funke",
+      avatar: "https://reqres.in/img/faces/9-image.jpg",
+    },
+    {
+      id: 10,
+      email: "byron.fields@reqres.in",
+      first_name: "Byron",
+      last_name: "Fields",
+      avatar: "https://reqres.in/img/faces/10-image.jpg",
+    },
+  ],
+  userDetail: null,
+  change: true,
+};
+const userReducer = (state = initialValues, action) => {
+  switch (action.type) {
+    case actionTypes.DELETE_USER:
+      let newUserDetail;
+      if (state.userDetail) {
+        if (action.id === state.userDetail.id) {
+          newUserDetail = null;
+        } else {
+          newUserDetail = state.userDetail;
+        }
+      } else {
+        newUserDetail = null;
+      }
+      return {
+        data: state.data.filter((user) => user.id !== action.id),
+        userDetail: newUserDetail,
+        change: state.change,
+      };
+    case actionTypes.SHOW_USER:
+      const detail = state.data.find((user) => user.id === action.id);
+      return {
+        data: state.data,
+        userDetail: detail,
+        change: !state.change,
+      };
+
+    default:
+      return state;
+  }
+};
+export default userReducer;
